+++ HardRayTrue/Viewport.cpp
@@ -1,12 +1,12 @@
 #include "Viewport.h"
 
 
-Vector3f Viewport::getPosFromCanvasPos(const Canvas& canvas, const sf::Vector2f &pos) const
+Vector3d Viewport::getPosFromCanvasPos(const Canvas& canvas, const sf::Vector2f &pos) const
 {
-	float x = (pos.x * size.x) / canvas.getSize().x; 
-	float y = (pos.y * size.y) / canvas.getSize().y;
-	float z = distanceToCamera; 
+	double x = (pos.x * size.x) / canvas.getSize().x; 
+	double y = (pos.y * size.y) / canvas.getSize().y;
+	double z = distanceToCamera; 
 
 
- 	return Vector3f(x, y, z);
+ 	return Vector3d(x, y, z);
 }
